Update balancedBinarySearchTree.py
The insert function initializes new Nodes as references in the tree and their children are initialized to None. Still, I need to find a way to point to the root to make it easier for users that way they don't have concerns about where they are looking for elements initially. Otherwise, all functions and traversals work as they should. Side Note: still need to implement unittest class for testing puroses.